Assume other compilers exist on Windows.
For now, we assume that if you're on Windows and not using MSVC,
you're using something that complies with GCC style arguments.

It also compiles the vulkan shared library as "vulkan" rather than
"vulkan-1". This is consistent with the Unix toolchain but may be
undesirable... some input on this would be nice.