Add MinGW support to build system.
Most of the issues regarding MinGW is that it's on Windows.
Most build systems simply assume that Windows means MSVC.
This is a rather poor assumption but generally easy to fix.

We build VKstatic.1.a and libvulkan.dll.a rather than libvulkan-1.dll.a

There's also a minor change to how paths are quoted. Some of the
quotations were treated poorly by cmake. When adding definitions,
quoting each invidividual definition is the best way to go. Otherwise,
CMake will add escape slashes which confuses GCC and probably even
VC++ in some case.